+++ tests/Storage/StorageFactoryTest.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 use Linna\Storage\MongoDbStorage;
 use Linna\Storage\MysqliStorage;
 use Linna\Storage\MysqlPdoStorage;
+use Linna\Storage\PostgresqlPdoStorage;
 use Linna\Storage\StorageFactory;
 use MongoDB\Client;
 use PHPUnit\Framework\TestCase;
@@ -35,6 +36,20 @@ class StorageFactoryTest extends TestCase
         $driver = (new StorageFactory('mysqlpdo', $options))->getConnection();
 
         $this->assertInstanceOf(MysqlPdoStorage::class, $driver);
+        $this->assertInstanceOf(\PDO::class, $driver->getResource());
+    }
+    
+    public function testCreatePostgresqlPdo()
+    {
+        $options = [
+            'dsn'      => $GLOBALS['pdo_pgsql_dsn'],
+            'user'     => $GLOBALS['pdo_pgsql_user'],
+            'password' => $GLOBALS['pdo_pgsql_password'],
+        ];
+
+        $driver = (new StorageFactory('pgsqlpdo', $options))->getConnection();
+
+        $this->assertInstanceOf(PostgresqlPdoStorage::class, $driver);
         $this->assertInstanceOf(\PDO::class, $driver->getResource());
     }